Started to re-implement the SVG part with d3.js
Don't know why I didn't think to start with D3 in the first place given that I already have a little experience with it. Apparently Snap did a great job in SEO and I'm easy to be distracted :-P

I took the chance to take an OOP approach this time.