Removes lib/usm and replaces it by the newly created package usm.io
lib/usm was moved to a separate package. This commit removes lib/usm from this package and replaces it by the newly created npm package.

This repo was renamed to usm.io-render-service so that the previous name usm.io can be used for the new repo.

The new repo usm.io is now developed and maintained at https://github.com/frederikheld/usm.io.git